+++ jest.config.js
@@ -1,3 +1,6 @@
-export const preset = "ts-jest";
-export const testEnvironment = "node";
-export const testMatch = ["**/__tests__/**/*.ts", "**/?(*.)+(spec|test).ts"];
+/* eslint-disable no-undef */
+module.exports = {
+  preset: "ts-jest",
+  testEnvironment: "node",
+  testMatch: ["**/__tests__/**/*.ts", "**/?(*.)+(spec|test).ts"],
+};