+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  
 ### Added
 
+- Implementing tests for Logger, Router and Server
+
 - Creating Simpler server README.md file.
 
 - Creating Simpler server response method.